pyproject.toml
[project]
name = "ai-video-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "google-genai>=1.25.0",
    "gradio>=5.35.0",
    "moviepy>=2.2.1",
    "openai>=1.82.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "sentencepiece>=0.2.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.3",
    "ultralytics>=8.3.163",
    "webrtcvad>=2.0.10",
]
